Updated all samples to use ReadAllEvents.ReadUntilEvent(TcpSocket spikeSafeSocket, SpikeSafeEvents code)

+++ src/Vektrex.SpikeSafe.CSharp.Samples/ApplicationSpecificExamples/FixedPulseCountUsingSoftwareTiming/FixedPulseCountUsingSoftwareTimingExample.cs
@@ -87,7 +87,7 @@ namespace Vektrex.SpikeSafe.CSharp.Samples.ApplicationSpecificExamples.FixedPuls
                 tcpSocket.SendScpiCommand("OUTP1 1");
 
                 // wait until Channel 1 is ready
-                ReadAllEvents.ReadUntilEvent(tcpSocket, (int)SpikeSafeEvents.CHANNEL_READY); // event 100 is "Channel Ready"
+                ReadAllEvents.ReadUntilEvent(tcpSocket, SpikeSafeEvents.CHANNEL_READY); // event 100 is "Channel Ready"
 
                 // pulsing starts before before getting Channel Ready message
                 // wait 30ms for getting ~10000 pulses
